@@ -3,8 +3,7 @@
   "version": "0.0.0",
   "private": true,
   "engines": {
-    "node": ">=6.5",
-    "npm": ">=3.10"
+    "node": "8.10.0"
   },
   "browserslist": [">1%", "last 4 versions", "Firefox ESR", "not ie < 9"],
   "dependencies": {
@@ -12,7 +11,7 @@
     "awesome-bootstrap-checkbox": "^1.0.1",
     "bluebird": "^3.5.1",
     "body-parser": "^1.18.2",
-    "bootstrap": "^4.1.1",
+    "bootstrap": "4.2.1",
     "bundle-loader": "^0.5.5",
     "classnames": "^2.2.5",
     "cookie-parser": "^1.4.3",
@@ -39,15 +38,15 @@
     "pretty-error": "^2.1.1",
     "prop-types": "^15.6.1",
     "query-string": "^5.0.1",
-    "react": "^16.1.1",
+    "react": "16.7.0",
     "react-cookie": "^1.0.5",
-    "react-dom": "^16.1.1",
+    "react-dom": "16.7.0",
     "react-google-maps": "^9.4.5",
     "react-redux": "^5.0.4",
-    "react-router": "^4.0.0",
-    "react-router-dom": "^4.0.0",
+    "react-router": "4.3.1",
+    "react-router-dom": "4.3.1",
     "react-sparklines": "^1.7.0",
-    "reactstrap": "^6.0.1",
+    "reactstrap": "7.1.0",
     "recharts": "^1.0.0-beta.10",
     "redux": "^3.6.0",
     "redux-logger": "^3.0.1",
