@@ -1,7 +1,7 @@
 /**
- * React Starter Kit (https://www.reactstarterkit.com/)
+ * Flatlogic Dashboards (https://flatlogic.com/admin-dashboards)
  *
- * Copyright © 2014-present Kriasoft, LLC. All rights reserved.
+ * Copyright © 2015-present Flatlogic, LLC. All rights reserved.
  *
  * This source code is licensed under the MIT license found in the
  * LICENSE.txt file in the root directory of this source tree.
@@ -19,6 +19,7 @@ import loadPrivacy from 'bundle-loader?lazy!../../pages/privacy';
 import loadProfile from 'bundle-loader?lazy!../../pages/profile';
 import loadTypography from 'bundle-loader?lazy!../../pages/typography';
 import loadTables from 'bundle-loader?lazy!../../pages/tables';
+import loadButtons from 'bundle-loader?lazy!../../pages/buttons';
 import loadNotifications from 'bundle-loader?lazy!../../pages/notifications';
 import loadCharts from 'bundle-loader?lazy!../../pages/charts';
 import loadIcons from 'bundle-loader?lazy!../../pages/icons';
@@ -40,6 +41,7 @@ const PrivacyBundle = Bundle.generateBundle(loadPrivacy);
 const ProfileBundle = Bundle.generateBundle(loadProfile);
 const TypographyBundle = Bundle.generateBundle(loadTypography);
 const TablesBundle = Bundle.generateBundle(loadTables);
+const ButtonsBundle = Bundle.generateBundle(loadButtons);
 const NotificationsBundle = Bundle.generateBundle(loadNotifications);
 const ChartsBundle = Bundle.generateBundle(loadCharts);
 const IconsBundle = Bundle.generateBundle(loadIcons);
@@ -78,6 +80,7 @@ class Layout extends React.Component {
               <Route path="/app/typography" exact component={TypographyBundle} />
               <Route path="/app/tables" exact component={TablesBundle} />
               <Route path="/app/notifications" exact component={NotificationsBundle} />
+              <Route path="/app/components/buttons" exact component={ButtonsBundle} />
               <Route path="/app/components/charts" exact component={ChartsBundle} />
               <Route path="/app/components/icons" exact component={IconsBundle} />
               <Route path="/app/components/maps" exact component={MapsBundle} />
